+++ src/api-utils/midddleware/webhookValidations.ts
@@ -4,8 +4,8 @@ import { isEmpty, isNaN } from 'lodash';
 
 import { TWO } from '../../games/consts/global';
 import { findGameTypeByClientSecret } from '../../models/GameType';
-import { isRequestFresh } from '../../modules/slack/utils';
 import { validateWebhookSignatures } from '../../utils/cryptography';
+import { isRequestFresh } from '../utils';
 
 export async function webhookValidation(request: Request, _h: ResponseToolkit) {
   if (!isEmpty(request.payload) && !Buffer.isBuffer(request.payload)) {
