+++ src/modules/dashboard/admin/adminRoutes/weaponAdminRoutes.ts
@@ -1,0 +1,64 @@
+import { CAPABILITIES } from '../../../../api-utils/interfaceAndTypes';
+import { getAuthUserMiddleware } from '../../../../api-utils/midddleware';
+import {
+  getWeaponsHandler,
+  getWeaponHandler,
+  upsertWeaponHandler,
+  deleteWeaponHandler,
+} from '../adminHandlers/weaponAdminHandlers';
+
+export const getWeaponsRoute = {
+  method: 'GET',
+  path: '/dashboard/admin/getWeapons',
+  options: {
+    description: 'Get list of weapons',
+    tags: ['api'],
+    bind: {
+      requiredCapabilities: [CAPABILITIES.ADMIN_ACTIONS],
+    },
+    pre: [getAuthUserMiddleware],
+  },
+  handler: getWeaponsHandler,
+};
+
+export const getWeaponByIdRoute = {
+  method: 'GET',
+  path: '/dashboard/admin/weapons/{weaponId}',
+  options: {
+    description: 'Get information on a specific weapon',
+    tags: ['api'],
+    bind: {
+      requiredCapabilities: [CAPABILITIES.ADMIN_ACTIONS],
+    },
+    pre: [getAuthUserMiddleware],
+  },
+  handler: getWeaponHandler,
+};
+
+export const upsertWeaponRoute = {
+  method: 'POST',
+  path: '/dashboard/admin/upsertWeapon',
+  options: {
+    description: 'Add or update weapon',
+    tags: ['api'],
+    bind: {
+      requiredCapabilities: [CAPABILITIES.ADMIN_ACTIONS],
+    },
+    pre: [getAuthUserMiddleware],
+  },
+  handler: upsertWeaponHandler,
+};
+
+export const deleteWeaponRoute = {
+  method: 'DELETE',
+  path: '/dashboard/admin/weapons/{itemId}',
+  options: {
+    description: 'Delete a weapon by its item id.',
+    tags: ['api'],
+    bind: {
+      requiredCapabilities: [CAPABILITIES.ADMIN_ACTIONS],
+    },
+    pre: [getAuthUserMiddleware],
+  },
+  handler: deleteWeaponHandler,
+};
