+++ src/routes.ts
@@ -1,8 +1,8 @@
 import type { ServerRoute } from '@hapi/hapi';
 
-import { adminRoutes } from './modules/dashboard/admin/adminRoutes';
-import { gameDevRoutes } from './modules/dashboard/gameDev/gameDevRoutes';
-import { gameDevWebhookRoutes } from './modules/gameDevWebhook/webhookRoutes';
+import { adminRoutes } from './modules/dashboard/admin/routes';
+import { gameDevRoutes } from './modules/dashboard/gameDev/routes';
+import { gameDevWebhookRoutes } from './modules/gameDevWebhook/routes';
 import { slackArenaRoutes } from './modules/slack/slackArenaRoute';
 import { slackRoutes } from './modules/slack/slackRoutes';
 import { slackTowerRoutes } from './modules/slack/slackTowerRoute';
