+++ test/test-utils.ts
@@ -15,7 +15,7 @@ declare module 'mocha' {
   }
 }
 
-// sequelize.options.logging = false;
+sequelize.options.logging = false;
 
 export const getCustomTestServer = () => {
   return new Server({
